akaros: Halt with mwait
This allows the guest to halt its dedicated cores in a deep enough state
for the other cores to reach the top turbo modes.

We halt with mwait instead of the usual halt instruction.  Both are
handled the same in software during vmexit handling.  When the VM has
dedicated host physical cores, the VMM will allow the guest to
halt/mwait directly (no vmexits).

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>