akaros: add support for running as a guest of the Akaros OS
This is a rebase, squash, and cleanup of our previous patchset on 5.1
with the latest commit a55137ab9f45 ("akaros: add an example for
building and deploying an app").

It's easier to rebase a single patch, especially when the final version
of it only touches a few files.

Fixes from v5.1 include:
- Renamed lapic_timer_frequency -> lapic_timer_period
- Rebased to 5.3

Signed-off-by: Barret Rhoden <brho@google.com>