akaros: add support for running as a guest of the Akaros OS
This is a rebase, squash, and cleanup of our previous patchset on 5.3,
plus whatever fixes are necessary in the rebase.

It's easier to rebase a single patch, especially when the final version
of it only touches a few files.

Fixes from v5.3 include:
- Turned on Xen configs, CONFIG_PARAVIRT_XXL silently depends on
CONFIG_XEN, otherwise you run into compilation issues with xen_iret.
- Turned on a bunch of configs for virtio and other things I want.  And
a few that cloud-hypervisor wants, like IOMMU and PCI_MSI...
- The real_mode trampoline was a pain in the ass.  We have no memory
below 1 MB, and the code wanted to make sure it could allocate something
from there.  We never actually tramp there, so allocating it anywhere
else should be OK.
- Disabled a bunch of cpu capabilities, which were triggering MSR
failures
- Made static networking in tc-sys.sh more robust.
- Exported kthread_create_on_cpu, which is used by the Akaros repo's
tests/linux/modules/.
- Rebased to 5.16

Signed-off-by: Barret Rhoden <brho@google.com>