akaros: rewrite TCZ installation
Previously we were setting up the initramfs so that it would install the
tinycore packages when the kernel booted.  That's a waste of boot time.

This commit changes the build script to extract the packages when we
build the initramfs.

It also sorts out the package dependencies for you.  Previously, you had
to find them yourself.  Now just list the TCZ package name you want.

This also adds a small framework to add custom binaries from your
machine to the initramfs.

The purpose of all of this is to make it easier to build simple images
that boot quickly.  It adds everything you need, and as little of what
you don't as possible.

Signed-off-by: Barret Rhoden <brho@google.com>