akaros: initrd: fix dangling symlink issue with chmod
Certain TC packages, such as iproute2, may bring in files with dangling
symlinks.  That might be due to some issue elsewhere in the scripts.
You'd get this error:

	chmod: cannot operate on dangling symlink 'tc_root/usr/local/bin/iptables-xml'

chmod will attempt to follow links if they are explicitly passed on the
command line.  When the link is dangling, chmod fails, and the scripts
abort.

The fix is to use a recursive chmod, instead of explicitly chmodding
files.  This also helps catch other files under u/l/b/ that were not
directly in bin/.

Signed-off-by: Barret Rhoden <brho@google.com>